refactor: LabeledUserDropdown ⇒ ParticipantDropdown
Refactoring this to prepare it as a shared component. Step 1: rename it
to something that better describes its purpose.
