+++ src/modules/Dashboard/components/TakeNowModal/ParticipantDropdown/ParticipantDropdown.styles.ts
