+++ src/modules/Dashboard/components/TakeNowModal/ParticipantDropdown/index.ts
@@ -1,0 +1,2 @@
+export * from './ParticipantDropdown';
+export * from './ParticipantDropdown.types';
