+++ src/modules/Dashboard/components/TakeNowModal/TakeNowModal.types.ts
@@ -1,6 +1,6 @@
 import { ModalProps } from 'shared/components';
 
-import { ParticipantDropdownOption } from './LabeledDropdown/LabeledUserDropdown.types';
+import { ParticipantDropdownOption } from './ParticipantDropdown';
 
 export type UseTakeNowModalProps = {
   dataTestId: string;
